requirements.txt
requests
gql
aiohttp
python-json-logger
varname
pytenable
